Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-jose
spacy
